Legal_Document_Analyzer_AI/requirements.txt
# Core Streamlit and Data Science Libraries
streamlit==1.49.1
pandas==2.2.3
#numpy==1.26.4
plotly==5.24.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# HTTP and Environment
requests==2.32.3
python-dotenv==1.0.1

# Development Tools
black==24.10.0
flake8==7.1.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
